+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## [Unreleased] (2025-07-20)
 
+### Features
+- Provide diagnostics for collectons settings files
+
 ### Fixes / Improvements
 - Make all file system operations async (for reducing waiting time before being able to continue with other actions for cases where a file system operation takes longer for some reason).
-- update all npm packages
-- bump Bruno CLI version used for executing tests via npx
+- Update all npm packages
+- Bump Bruno CLI version used for executing tests via npx to '2.7.0'
 
 ## [0.13.1] (2025-07-16)